refactor: 코드 리팩토링 (#11)
