+++ src/atoms/resultState.ts
@@ -1,15 +1,14 @@
 import {atom, RecoilState} from 'recoil'
 import {ResultData} from '../types/ResultData'
 import {Result} from '../types/Result'
-import {Course} from '../types/Course'
-import {Culture} from '../types/Culture'
+import {Course, CourseCulture} from '../types/Course'
 
 const initState: ResultData = {
   advanced: {} as Result<Course>,
   optional: {} as Result<Course>,
-  basic: {} as Result<Culture>,
-  core: {} as Result<Culture>,
-  balance: {} as Result<Culture>
+  basic: {} as Result<CourseCulture>,
+  core: {} as Result<CourseCulture>,
+  balance: {} as Result<CourseCulture>
 }
 
 const resultState: RecoilState<ResultData> = atom({
