+++ src/pages/result/ResultPage.tsx
@@ -2,8 +2,7 @@ import {useState} from 'react'
 import ResultCultureItem from '../../components/result/ResultCultureItem'
 import ResultDetailItem from '../../components/result/ResultDetailItem'
 import ResultItem from '../../components/result/ResultItem'
-import {Category} from '../../types/Course'
-import {Domain} from '../../types/Culture'
+import {Category, Domain} from '../../types/Course'
 import ModalContainer from '../../components/result/ModalContainer'
 import ResultContainer from '../../components/result/ResultContainer'
