--- src/types/Culture.ts
@@ -1,7 +1,0 @@
-export interface Culture {
-  subDomain: string
-  subDomainName: string
-  completed: boolean
-}
-
-export type Domain = 'BASIC' | 'CORE' | 'BALANCE'
